# everyfoody/swagger.json
{
  "swagger": "2.0",
  "schemes": [
    "http",
    "https"
  ],
  "host": "220.230.114.8:3444",
  "basePath": "/",
  "info":
  {
    "description": "Dream comes truck!!<!-- ReDoc-Inject: <security-definitions> -->\n",
    "version": "1.0.0",
    "title": "EveryFoody",
    "x-logo":
    {
      "url": "https://s3.ap-northeast-2.amazonaws.com/everyfoody/everyfoody.png"
    }
  },
...
  "securityDefinitions":
  {
    "petstore_auth":
    {
      "description": "Get access to data while protecting your account credentials.\nOAuth2 is also a safer and more secure way to give you access.\n",
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/api/oauth/dialog",
      "flow": "implicit",
      "scopes":
      {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key":
    {
      "description": "For this sample, you can use the api key `special-key` to test the authorization filters.\n",
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "paths":
...
    "/signup/customer":
    {
      "post":
      {
        "tags": [
          "sign"
        ],
        "summary": "이용자 회원가입",
        "description": "Facebook / Kakao로 이용자 회원가입을 합니다.",
        "operationId": "customer sign up",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
        {
          "in": "body",
          "name": "body",
          "description": "유저 회원가입을 하기 위한 email, uid, name, category 전달",
          "required": true,
          "schema":
          {
            "type": "object",
            "properties":
            {
              "email":
              {
                "description": "페이스북 or 카카오톡 id",
                "type": "string",
                "minLength": 1,
                "example": "everyfoody@gmail.com"
              },
              "uid":
              {
                "description": "발급된 uid",
                "type": "string",
                "minLength": 1,
                "example": "10985781390487"
              },
              "name":
              {
                "description": "유저가 적은 이름",
                "type": "string",
                "minLength": 1
              },
              "category":
              {
                "description": "Facebook/Kakao 구분, Facebook : 1 Kakao : 2",
                "type": "integer",
                "minLength": 1,
                "example": 1
              }
            }
          }
        }],
        "responses":
        {
          "201":
          {
            "description": "이용자 회원가입 성공",
            "schema":
            {
              "type": "object",
              "properties":
              {
                "status":
                {
                  "type": "string",
                  "description": "성공 알림",
                  "example": "success"
                },
                "msg":
                {
                  "type": "string",
                  "description": "회원가입 성공 알림 메세지",
                  "example": "successful customer signup"
                }
              }
            }
          }
        }
      }
    },
    "/signup/owner":
    {
      "post":
      {
        "tags": [
          "sign"
        ],
        "summary": "사업자 회원가입",
        "description": "Facebook / Kakao로 사업자 회원가입을 합니다.",
        "operationId": "owner signup",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
        {
          "in": "body",
          "name": "body",
          "description": "사업자 회원가입을 하기 위한 정보 전달",
          "required": true,
          "schema":
          {
            "type": "object",
            "properties":
            {
              "email":
              {
                "description": "페이스북 or 카카오톡 id",
                "type": "string",
                "minLength": 1,
                "example": "everyfoody@gmail.com"
              },
              "uid":
              {
                "description": "발급된 uid",
                "type": "string",
                "minLength": 1,
                "example": "10985781390487"
              },
              "name":
              {
                "description": "사업자가 적은 이름",
                "type": "string",
                "minLength": 1,
                "example": "김연태"
              },
              "category":
              {
                "description": "Facebook/Kakao 구분, Facebook : 1 Kakao : 2",
                "type": "integer",
                "minLength": 1,
                "example": 1
              }
            }
          }
        }],
        "responses":
        {
          "201":
          {
            "description": "사업자 회원가입 성공",
            "schema":
            {
              "type": "object",
              "properties":
              {
                "status":
                {
                  "type": "string",
                  "description": "성공 알림",
                  "example": "success"
                },
                "msg":
                {
                  "type": "string",
                  "description": "사업자 회원가입 성공알림 메세지",
                  "example": "successful owner signup"
                }
              }
            }
          }
        }
      }
    },
    "/main/lists/:location/:latitude/:longitude":
    {
      "get":
      {
        "tags": [
          "main"
        ],
        "summary": "위치별 푸드트럭리스트",
        "description": "GPS기반 구별 푸드트럭 리스트 정보",
        "operationId": "basicInfo",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
        {
          "in": "header",
          "name": "JWT token",
          "description": "사용자 인증을 위한 token",
          "required": true,
          "type": "string"
        },
        {
          "in": "query",
          "name": "location",
          "description": "위치 구분값, 1 : 강서양천영등포구로, 2 : 은평마포서대문, 3 : 종로중구용산, 4 : 도봉강북성북노원, 5 : 동대문중랑성동광진, 6 : 동작관악금천, 7 : 서초강남, 8 : 강동송파",
          "required": true,
          "type": "integer"
        },
        {
          "in": "query",
          "name": "latitude",
          "description": "사용자의 위도",
          "required": true,
          "type": "double"
        },
        {
          "in": "query",
          "name": "longitude",
          "description": "사용자의 경도",
          "required": true,
          "type": "double"
        }],
        "responses":
        {
          "200":
          {
            "description": "가까운 위치별로 10개씩 제공된 푸드트럭 정보",
            "schema":
            {
              "type": "object",
              "properties":
              {
                "status":
                {
                  "type": "string",
                  "description": "성공 알림",
                  "example": "success"
                },
                "data":
                {
                  "type": "array",
                  "items":
                  {
                    "type": "object",
                    "properties":
                    {
                      "storeID":
                      {
                        "type": "integer",
                        "description": "푸드트럭 인덱스용 푸드트럭ID",
                        "example": 13
                      },
                      "storeName":
                      {
                        "type": "string",
                        "description": "푸드트럭 이름",
                        "example": "김준영의 푸드트럭"
                      },
                      "storeImage":
                      {
                        "type": "string",
                        "description": "푸드트럭 사진 URL",
                        "example": "image url"
                      },
                      "reservationCount":
                      {
                        "type": "integer",
                        "description": "푸드트럭 예약자 현황",
                        "example": 8
                      },
                      "storeLocation":
                      {
                        "type": "string",
                        "description": "푸드트럭 위치",
                        "example": "서울시 마포구 상수동 65"
                      },
                      "storeDistance":
                      {
                        "type": "double",
                        "description": "푸드트럭까지 거리. 단위 : m or Km",
                        "example": 980
                      },
                      "storeDistanceUnit":
                      {
                        "type": "string",
                        "description": "푸드트럭까지 거리단위, 'm' or 'Km'",
                        "example": "m"
                      }
                    }
                  }
                },
                "msg":
                {
                  "type": "string",
                  "description": "푸드트럭정보 로딩 성공 메세지",
                  "example": "successful load store list"
                }
              }
            }
          },
          "500":
          {
            "description": "get connection error \nuser authorization error \nget store data error",
            "schema":
            {
              "$ref": "#/definitions/failMessage"
            }
          }
        }
      }
    },
...
    "/store/location/:storeID":
    {
      "get":
      {
        "tags": [
          "store information"
        ],
        "summary": "푸드트럭 위치정보",
        "description": "지도에 표시할 푸드트럭 위치정보",
        "operationId": "location",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
        {
          "in": "header",
          "name": "JWT token",
          "description": "사용자 인증을 위한 token",
          "required": true,
          "type": "string"
        },
        {
          "in": "query",
          "name": "storeID",
          "description": "가게 id",
          "required": true,
          "type": "integer"
        }],
        "responses":
        {
          "200":
          {
            "description": "가게 id에 해당하는 상세정보",
            "schema":
            {
              "type": "object",
              "properties":
              {
                "status":
                {
                  "type": "string",
                  "description": "성공 알림",
                  "example": "success"
                },
                "data":
                {
                  "type": "object",
                  "properties":
                  {
                    "storeID":
                    {
                      "type": "integer",
                      "description": "푸드트럭 인덱스용 푸드트럭ID",
                      "example": 13
                    },
                    "storeLatitude":
                    {
                      "type": "double",
                      "description": "푸드트럭 위도",
                      "example": 44.012380
                    },
                    "storeLongitude":
                    {
                      "type": "double",
                      "description": "푸드트럭 경도",
                      "example": 45.968741
                    },
                    "reservationCount":
                    {
                      "type": "integer",
                      "description": "푸드트럭 예약자 수",
                      "example": 3
                    }
                  }
                }
              },
              "msg":
              {
                "type": "string",
                "description": "푸드트럭 위치정보 로딩 성공 메세지",
                "example": "successful load location data"
              }
            }
          }
        },
        "500":
        {
          "description": "get connection error \nuser authorization error \nget store location error",
          "schema":
          {
            "$ref": "#/definitions/failMessage"
          }
        }
      }
    }
  },
  "definitions":
...
    "failMessage":
    {
      "type": "object",
      "properties":
      {
        "status":
        {
          "type": "string",
          "example": "fail"
        },
        "message":
        {
          "type": "string",
          "example": "fail reason"
        }
      }
    },
...
  }
}